Move default burshes creation and add built-in eraser brushes
