+++ Ciallo/Localization/TranslatableStrings.csv
@@ -125,4 +125,5 @@ Marker radius,Marker radius,标记半径,標記半徑,マーカー半径,마커 
 Show reference wireframe,Show reference wireframe,显示参考线框,顯示參考線框,参照ワイヤーフレームを表示,참조 와이어프레임 표시,Referenz-Drahtgitter anzeigen,Afficher le fil de référence
 Eraser mode,Eraser mode,橡皮擦模式,橡皮擦模式,消しゴムモード,지우개 모드,Radiergummi-Modus,Mode gomme
 Eraser,eraser,橡皮擦,橡皮擦,消しゴム,지우개,Radiergummi,Gomme
+Soft eraser,Soft eraser,软橡皮擦,軟橡皮擦,ソフト消しゴム,부드러운 지우개,weicher Radiergummi,Gomme douce
 Mapping pen,Mapping pen,丸笔,丸笔,丸ペン,맵핑 펜,Mapping-Stift,Stylo de cartographie
